[fix] multiple qr30 payment

+++ static/src/app/screens/payment_screen/payment_screen.js
@@ -1,5 +1,7 @@
 import { PaymentScreen } from "@point_of_sale/app/screens/payment_screen/payment_screen";
 import { patch } from "@web/core/utils/patch";
+import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
+import { _t } from "@web/core/l10n/translation";
 
 patch(PaymentScreen.prototype, {
   setup() {
@@ -29,42 +31,77 @@ patch(PaymentScreen.prototype, {
     this.env.services.bus_service.start();
   },
   async addNewPaymentLine(paymentMethod) {
-    var result = await this.pos.data.call(
-      "pos.payment.method",
-      "get_scb_qr_fees",
-      [[paymentMethod.id], this.currentOrder.get_due()]
-    );
-    if (result && result.fix_payment_product_id && result.fix_payment_fees) {
-      let product = this.pos.models["product.product"].get(
-        result.fix_payment_product_id
-      );
-      if (!product) {
-        product = await this.pos.data.searchRead("product.product", [
-          ["id", "=", result.fix_payment_product_id],
-        ]);
-        product = product.length > 0 && product[0];
-      }
-      let line = this.currentOrder.lines.find(
-        (line) => line.product_id.id === product.id
-      );
-      if (line) {
-        line.set_unit_price(result.fix_payment_fees);
-      } else {
-        await this.pos.addLineToCurrentOrder(
-          {
-            product_id: product,
-            price_unit: result.fix_payment_fees,
-          },
-          {}
-        );
-      }
-      this.currentOrder.setFixPaymentProductId(result.fix_payment_product_id);
+    if (paymentMethod.qr_code_method != "qr30")
+      return await super.addNewPaymentLine(paymentMethod);
+
+    if (paymentMethod.qr30_payment_fee <= 0)
+      return await super.addNewPaymentLine(paymentMethod);
+
+    const product_id = paymentMethod._raw.qr30_payment_fee_product_id;
+    if (!product_id) {
+      this.dialog.add(AlertDialog, {
+        title: _t("Error"),
+        body: _t(
+          "Invalid payment method configuration: payment fee product is not set."
+        ),
+      });
+      return;
     }
-    return await super.addNewPaymentLine(paymentMethod);
+
+    if (this.currentOrder.getFixPaymentProductId()) {
+      this.dialog.add(AlertDialog, {
+        title: _t("Error"),
+        body: _t("Only one QR30 payment is allowed."),
+      });
+      return;
+    }
+
+    let product = this.pos.models["product.product"].get(product_id);
+    if (!product) {
+      product = await this.pos.data.searchRead("product.product", [
+        ["id", "=", product_id],
+      ]);
+      // get first product
+      product = product.length > 0 && product[0];
+    }
+    if (!product) {
+      this.dialog.add(AlertDialog, {
+        title: _t("Error"),
+        body: _t(
+          "Invalid payment method configuration: payment fee product not found."
+        ),
+      });
+      return;
+    }
+
+    let line = this.currentOrder.lines.find(
+      (line) => line.product_id.id === product.id
+    );
+
+    if (line) {
+      console.log(line.get_quantity());
+      line.set_quantity(line.get_quantity() + 1, true);
+    } else {
+      line = await this.pos.addLineToCurrentOrder(
+        {
+          product_id: product,
+          price_unit: paymentMethod.qr30_payment_fee,
+        },
+        {}
+      );
+    }
+
+    const result = await super.addNewPaymentLine(paymentMethod);
+    if (!result) {
+      line.delete();
+    } else {
+      this.currentOrder.setFixPaymentProductId(product_id);
+    }
+    return result;
   },
   deletePaymentLine(uuid) {
     const line = this.paymentLines.find((line) => line.uuid === uuid);
-    if (line.payment_method_id.payment_method_type === "qr_code") {
+    if (line.payment_method_id.qr_code_method === "qr30") {
       this.currentOrder.setShowQR(false);
       if (this.currentOrder.getFixPaymentProductId()) {
         let line = this.currentOrder.lines.find(
@@ -72,7 +109,11 @@ patch(PaymentScreen.prototype, {
             line.product_id.id === this.currentOrder.getFixPaymentProductId()
         );
         if (line) {
-          line.delete();
+          if (line.get_quantity() > 1) {
+            line.set_quantity(line.get_quantity() - 1, true);
+          } else {
+            line.delete();
+          }
           this.currentOrder.setFixPaymentProductId(false);
         }
       }
